Cleanup logic in firecheckout compatibility mixin
